Refactor parser
The parser has been refactored to return a ParseToken after it parses a
stream of Tokens. This is because Tokens contain other types of tokens
such as parentheses, commas, and other syntactical elements that are not
needed by the interpreter. A ParseToken does not contain these elements,
so it makes it simpler to cover all bases later on in
compilation/execution.