Variables can store values in memory
There are several tests that make sure that the variable parsing and
interpreting code works properly. The variants for Expr and Stmt that
deal with variables now use a struct called VarInfo which stores the
name, row and column of the variable being used. This makes it easier to
deal with variables in the interpreter because it means that we don't
have to match ParseToken every single time we want to get data out of a
variable.