Add rationale for lexer and parser refactor
The lexer and parser will be refactored to make it simpler to perform
operations on tokens. Currently, everything is stored in one token, so
it is hard to perform match operations, and loop conditions. After the
redesign, data and token type will be stored separately in a struct so
it will be simpler to perform these operations.